Initial Launcher GUI rewrite
The goal is to make this more configurable.
The long term goal for the starloader-project would be to support binary
patches. I've been privately commited to the neat little project that is
called "SML0" whose goal is to recompile galimulator. Right now it's
miles away from actually doing this, but I've also made signficiant
progress towards it (mostly with my involvement with several fernflower
forks and the QuiltMC community)