Do not acquire extension classloaders as locks
Previously if a class was loaded by a Thread X and another class which was
present in the extension classloader was loaded in another thread Y,
the thread X would have a lock on the root classloader and thread Y
a lock on the extension classloader. If the extension whom the extension
classloader belonged defined an ASM Transformer that modified the class
loaded by Thread X a deadlock would occur as soon as the Transformer
would try to load additional classes.

The Transformer would attempt to obtain a lock on the extension classloader
since it is not parallel capable, where as the extension classloader would
try to obtain a lock on the root classloader.

There is the possibility of such a change backfiring and causing issues
with thread safety. But until then - this should be an adequate solution,
even if it isn't an outright fix - that would require rethinking how
ASM Transformers work or at the very least how the class transformation
logic behaves. That however would lead to even less stable classloading
at least for the time being and was as such deemed to not be adequate.