Fix issues with IDE launching
It's a shame that maven does not share the same support when working with
scopes as gradle does. As such we were ultimately forced to expose the
internal (technically implementation-specific) micromixin dependency, so
eclipse can pick it up.

This didn't used to be a problem back when we were using a fat jar.
This isn't really a problem even in the current state as you can still
use the `shaded` classifier, but it'd be flawed nonetheless.

---

Note (9th Jan 2024): In the "spongeian" release branch, the
implementation-specific sponge transformer classes are wholly exposed.
We do not plan on changing that anytime soon. People should depend on
the micromixin branch whenver appropriate.