Get rid of GSON
GSON was previously used by sponge's mixin framework; but since
micromixin is not dependent on it GSON can be removed from the
micromixin variant of SLL

---

Note: This commit was originally made for the launcher-micromixin
variant of SLL, but in the launcher-sponge variant of SLL GSON is
still needed as a transitive dependency of micromixin. It will thus
continue to be present at runtime within launcher-sponge (but not
launcher-micromixin!!).

This cherry-pick also includes the extension loading hotfix
commit b122c6a26654aa69f27e3709cf82023ea1f0105e - but it surprises
me that everything works as it should in this environment
if I'll be completely honest.