Fixed dependencies not being resolved properly
Extension classloaders were stored in a weak map.
However, the issue was that the key was basically guaranteed to be gone near-instantly,
thus causing the entry to be deleted off the map and the ExtensionManager believing
that dependents don't have the needed dependencies available.

That being said, this bug existed since day 1 and I am extremely surprised
I only discovered this bug now. Could have been JVM-specific? But even that
is a bit of a stretch to the point that I'd blame a bad commit from the previous
days, but that is extremely unlikely to be the cause. Oh well.