Fix getResourceAsStreamWithChildren querying parent classloaders
The method will now query the classloader itself and it's children.

This change fixes an issue where the development environment wouldn't
properly apply mixins when launching through eclipse's launch system.
This mostly affects environments with asymmetric deobfuscation mappings,
as the resource on the classpath will not be remapped to the common
deobfuscation mapping, causing mixins to fail to apply at times.